Enhances test output and refactors core logic
Improves readability of test output by adding visual enhancements.

Refactors test execution engine and YAML parsing into a `core` directory, improving project structure and separation of concerns.

Removes `chalk` dependency from browser engine and core, simplifying the codebase and reducing dependencies.
